Move SpiralIterator2d file
